Backend/test_crop_detection.py
"""
Test the crop detection model integration
"""

import sys
from pathlib import Path

# Test 1: Check if model file exists
print("\n🧪 Testing Crop Detection Model Integration")
print("=" * 60)

model_path = Path("crop_detection_model.pt")
if model_path.exists():
    print(f"✅ Model file found: {model_path}")
    print(f"   Size: {model_path.stat().st_size / 1024 / 1024:.2f} MB")
else:
    print(f"❌ Model file not found: {model_path}")
    sys.exit(1)

# Test 2: Load the model
print("\n📦 Loading YOLOv8 model...")
try:
    from ultralytics import YOLO
    model = YOLO(str(model_path))
    print(f"✅ Model loaded successfully")
    print(f"   Model type: {type(model).__name__}")
except Exception as e:
    print(f"❌ Failed to load model: {e}")
    sys.exit(1)

# Test 3: Check model classes
print("\n🏷️  Model Classes:")
if hasattr(model, 'names'):
    classes = model.names
    print(f"   Total classes: {len(classes)}")
    print(f"   Sample classes: {list(classes.values())[:10]}")
else:
    print("   ⚠️  No class names available")

# Test 4: Test StorageGuardAgent integration
print("\n🤖 Testing StorageGuardAgent integration...")
try:
    from app.agents.storage_guard import StorageGuardAgent
    
    agent = StorageGuardAgent()
    print(f"✅ StorageGuardAgent initialized")
    
    model_info = agent.get_model_info()
    print(f"\n📊 Agent Model Info:")
    print(f"   Model Type: {model_info['model_type']}")
    print(f"   Model Path: {model_info['model_path']}")
    print(f"   Is Mock: {model_info['is_mock']}")
    print(f"   Status: {model_info['status']}")
    print(f"   Num Classes: {model_info.get('num_classes', 'N/A')}")
    
    if model_info['is_mock']:
        print("\n⚠️  WARNING: Agent is using mock model!")
        print("   Check if crop_detection_model.pt is in the correct location")
    else:
        print("\n✅ Agent is using the crop detection model!")
        
except Exception as e:
    print(f"❌ Failed to initialize StorageGuardAgent: {e}")
    import traceback
    traceback.print_exc()
    sys.exit(1)

# Test 5: Test with a sample image (if available)
print("\n🖼️  Testing image analysis...")
try:
    import requests
    from io import BytesIO
    from PIL import Image
    import numpy as np
    
    # Use a sample tomato image from the internet
    test_image_url = "https://images.unsplash.com/photo-1546470427-227eb1e2a2c2?w=400"
    
    print(f"   Downloading test image...")
    response = requests.get(test_image_url, timeout=10)
    
    if response.status_code == 200:
        image_bytes = response.content
        
        # Test with StorageGuardAgent
        report = agent.analyze_image(image_bytes)
        
        print(f"\n📋 Analysis Report:")
        print(f"   Quality Grade: {report.overall_quality}")
        print(f"   Shelf Life: {report.shelf_life_days} days")
        print(f"   Defects Found: {report.defects_found}")
        
        if hasattr(report, 'crop_detected') and report.crop_detected:
            print(f"   🌾 Crop Detected: {report.crop_detected}")
            print(f"   Confidence: {report.crop_confidence:.2%}")
        else:
            print(f"   🌾 Crop Detected: Not identified")
        
        print("\n✅ Image analysis working!")
    else:
        print(f"   ⚠️  Could not download test image (status: {response.status_code})")
        
except Exception as e:
    print(f"   ⚠️  Image test skipped: {e}")

print("\n" + "=" * 60)
print("✅ All tests passed! Crop detection model is ready.")
print("\n📝 Next steps:")
print("   1. Upload crop images via the frontend")
print("   2. Check quality analysis results")
print("   3. Verify crop detection in the response")
