Backend/requirements.txt
absl-py==2.3.1
accelerate==1.10.1
acres==0.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anthropic==0.69.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
astunparse==1.6.3
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.1.3
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.3
crypto==1.4.1
cryptography==46.0.1
cycler==0.12.1
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
effdet==0.4.1
email-validator==2.3.0
etelemetry==0.3.1
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.1
filelock==3.13.1
PyMuPDF==1.24.10
flatbuffers==25.9.23
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2024.6.1
gast==0.6.0
GeoAlchemy2==0.18.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.1
humanfriendly==10.0
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
PyJWT==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
keras==3.11.3
kiwisolver==1.4.9
langchain-core==0.3.76
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.31
libclang==18.1.1
loguru==0.7.3
looseversion==1.3.0
lxml==6.0.2
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
mmdet==3.3.0
mpmath==1.3.0
multidict==6.6.4
Naked==0.1.32
namex==0.1.0
networkx==3.3
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
omegaconf==2.3.0
onnx==1.19.0
onnxruntime-gpu==1.23.0
openai==2.1.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.2
passlib==1.7.4
pathlib==1.0.1
peft==0.17.1
pillow==11.3.0
platformdirs==4.4.0
polars==1.34.0
polars-runtime-32==1.34.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
prov==2.1.1
psutil==7.1.0
psycopg2-binary==2.9.10
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.10
pycparser==2.23
pydantic==2.11.9
pydantic-extra-types==2.10.5
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydot==4.0.1
Pygments==2.19.2
pymongo==4.15.3
pyparsing==3.2.5
pyreadline3==3.5.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.3
rdflib==7.2.1
regex==2025.9.18
reportlab==4.4.4
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
sendgrid==6.12.5
sentry-sdk==2.39.0
shapely==2.1.2
shellescape==3.8.1
shellingham==1.5.4
simplejson==3.20.2
simsimd==6.5.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
SQLAlchemy-Utils==0.42.0
starlette==0.48.0
stringzilla==4.0.14
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminaltables==3.1.10
threadpoolctl==3.6.0
timm==1.0.20
tokenizers==0.22.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.2
tqdm==4.67.1
traits==7.0.2
transformers==4.56.2
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
ultralytics==8.3.214
ultralytics-thop==2.0.17
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
wandb==0.22.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
win32_setctime==1.2.0
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
zstandard==0.25.0
